src/requirements.txt
Flask
gunicorn
PyYAML
psycopg2
itsdangerous
requests
SQLAlchemy
Flask-WTF
Werkzeug
tox
Flask-SQLAlchemy
connexion[swagger-ui]
flask-debugtoolbar-sqlalchemy
PyJWT
python-logstash
Flask-Mail
uuid
Pillow